sasl!
svn path=/trunk/kdesupport/qca/; revision=245060

+++ base64.cpp
@@ -115,10 +115,10 @@ QByteArray Base64::decode(const QByteArray &s)
 	c = d = 0;
 
 	for( i = 0; i < len; i += 4 ) {
-		a = tbl[s[i]];
-		b = tbl[s[i + 1]];
-		c = tbl[s[i + 2]];
-		d = tbl[s[i + 3]];
+		a = tbl[(int)s[i]];
+		b = tbl[(int)s[i + 1]];
+		c = tbl[(int)s[i + 2]];
+		d = tbl[(int)s[i + 3]];
 		if((a == 64 || b == 64) || (a < 0 || b < 0 || c < 0 || d < 0)) {
 			p.resize(0);
 			return p;
